Added python C extension for running the simulator

+++ src/minion.hpp
@@ -133,13 +133,13 @@ public:
   }
 
   void add_deathrattle_microbots(int n=3) {
-    this->deathrattle_microbots = min(this->deathrattle_microbots+n,7);
+    this->deathrattle_microbots = min(static_cast<int>(this->deathrattle_microbots)+n,7);
   }
   void add_deathrattle_golden_microbots(int n=3) {
-    this->deathrattle_golden_microbots = min(this->deathrattle_golden_microbots+n,7);
+    this->deathrattle_golden_microbots = min(static_cast<int>(this->deathrattle_golden_microbots)+n,7);
   }
   void add_deathrattle_plants(int n=2) {
-    this->deathrattle_plants = min(this->deathrattle_plants+n,7);
+    this->deathrattle_plants = min(static_cast<int>(this->deathrattle_plants)+n,7);
   }
 
   // events
